refactor(1175): Découpage des test de la fiche jeune dans des fichiers dédiés
